修改手机号、修改笔记、修改班群、修改分数小数点

+++ app/src/main/java/com/bll/lnkteacher/mvp/model/testpaper/AnalyseItem.java
@@ -13,7 +13,7 @@ public class AnalyseItem implements Serializable {
     public int rightNum;//正确人数
     public int wrongNum;//错误人数
     public double totalScore;//统计总分
-    public int totalLabel;//每题标准总分
+    public double totalLabel;//每题标准总分
     public double averageScore;//统计平均分
     public double scoreRate;//得分率
     public List<UserBean> rightStudents=new ArrayList<>();//题目正确学生
@@ -23,9 +23,9 @@ public class AnalyseItem implements Serializable {
     public static class UserBean{
         public int userId;
         public String name;
-        public int score;
+        public double score;
 
-        public UserBean(int userId,String name,int score){
+        public UserBean(int userId,String name,double score){
             this.userId=userId;
             this.name=name;
             this.score=score;
