+++ app/src/main/java/com/bll/lnkteacher/mvp/model/testpaper/ScoreItem.java
@@ -11,6 +11,6 @@ public class ScoreItem implements Serializable {
     public String score="0";
     public int sort;
     public int result;//0错1对
-    public int label;//题目标准分数
+    public double label;//题目标准分数
     public List<ScoreItem> childScores=new ArrayList<>();
 }
